Refactor asset into modular components
