+++ portals/knowledge-management/src/Components/Assets/AssetFilterTabs.vue
@@ -1,0 +1,72 @@
+<!--
+<COPYRIGHT>
+
+    Copyright © 2016-2026, Canyon GBS LLC. All rights reserved.
+
+    Aiding App™ is licensed under the Elastic License 2.0. For more details,
+    see <https://github.com/canyongbs/aidingapp/blob/main/LICENSE.>
+
+    Notice:
+
+    - You may not provide the software to third parties as a hosted or managed
+      service, where the service provides users with access to any substantial set of
+      the features or functionality of the software.
+    - You may not move, change, disable, or circumvent the license key functionality
+      in the software, and you may not remove or obscure any functionality in the
+      software that is protected by the license key.
+    - You may not alter, remove, or obscure any licensing, copyright, or other notices
+      of the licensor in the software. Any use of the licensor's trademarks is subject
+      to applicable law.
+    - Canyon GBS LLC respects the intellectual property rights of others and expects the
+      same in return. Canyon GBS™ and Aiding App™ are registered trademarks of
+      Canyon GBS LLC, and we are committed to enforcing and protecting our trademarks
+      vigorously.
+    - The software solution, including services, infrastructure, and code, is offered as a
+      Software as a Service (SaaS) by Canyon GBS LLC.
+    - Use of this software implies agreement to the license terms and conditions as stated
+      in the Elastic License 2.0.
+
+    For more information or inquiries please visit our website at
+    <https://www.canyongbs.com> or contact us via email at legal@canyongbs.com.
+
+</COPYRIGHT>
+-->
+<script setup>
+    defineProps({
+        tabs: {
+            type: Array,
+            required: true,
+        },
+        modelValue: {
+            type: String,
+            required: true,
+        },
+    });
+
+    const emit = defineEmits(['update:modelValue']);
+</script>
+
+<template>
+    <div class="border-b border-gray-200 mb-2">
+        <ul class="flex flex-wrap -mb-px" role="tablist" aria-label="Asset filter tabs">
+            <li v-for="tab in tabs" :key="tab.key" role="presentation">
+                <button
+                    type="button"
+                    role="tab"
+                    :id="`assets-tab-${tab.key}`"
+                    :aria-selected="modelValue === tab.key"
+                    :aria-controls="`assets-panel-${tab.key}`"
+                    @click="emit('update:modelValue', tab.key)"
+                    :class="[
+                        'inline-flex items-center gap-2 px-4 py-3 text-sm font-medium border-b-2 transition-colors whitespace-nowrap',
+                        modelValue === tab.key
+                            ? 'border-[rgba(var(--primary-500),1)] text-[rgba(var(--primary-600),1)]'
+                            : 'border-transparent text-gray-500 hover:text-gray-700 hover:border-gray-300',
+                    ]"
+                >
+                    {{ tab.label }}
+                </button>
+            </li>
+        </ul>
+    </div>
+</template>
